+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Codemod now replaces top-level API when they have args, such as `z.string().url("Please enter a valid URL")` => `z.url("Please enter a valid URL")`. ([#18](https://github.com/nicoespeon/zod-v3-to-v4/pull/18))
 
+### Fixed
+
+- The codemod was a bit too greedy when migrating `z.record()` patterns and would sometimes add incorrect `z.string()` params to it's parent call expression (e.g., `z.array(z.record(…))` -> `z.array(z.string(), z.record(…))`). Not anymore! ([#17](https://github.com/nicoespeon/zod-v3-to-v4/pull/17))
+
 ## [1.2.0] - 2025-08-13
 
 ## Changed